% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ucihar}
\alias{ucihar}
\title{Human Activity Recognition Using Smartphones}
\format{An object of class \code{data.frame} with 10299 rows and 68 columns.}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Human+Activity+Recognition+Using+Smartphones}
}
\usage{
ucihar
}
\description{
Human activity data hosted at the UCI Machine Learning Repository.
}
\keyword{datasets}

