% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ucihar_avgs}
\alias{ucihar_avgs}
\title{Averages for Human Activity Recognition Using Smartphones}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 180 rows and 68 columns.}
\usage{
ucihar_avgs
}
\description{
Averages for Human Activity Recognition Using Smartphones
}
\keyword{datasets}

