% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ucihar_simple}
\alias{ucihar_simple}
\title{Mean and standard deviation measurements from Human Activity Recognition Using Smartphones data set}
\format{An object of class \code{data.frame} with 10299 rows and 66 columns.}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Human+Activity+Recognition+Using+Smartphones}
}
\usage{
ucihar_simple
}
\description{
Human activity data hosted at the UCI Machine Learning Repository.
}
\keyword{datasets}

