% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ucihar_training}
\alias{ucihar_training}
\title{Training data from Human Activity Recognition Using Smartphones data set}
\format{An object of class \code{data.frame} with 7352 rows and 564 columns.}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Human+Activity+Recognition+Using+Smartphones}
}
\usage{
ucihar_training
}
\description{
Human activity data hosted at the UCI Machine Learning Repository.
}
\keyword{datasets}

